tsql-core/src/test/java/com/xy/xsql/block/tsql/core/expression/AtTimeConverterTest.java
package com.xy.xsql.block.tsql.core.expression;

import com.xy.xsql.block.core.ReferenceBlockPrinter;
import com.xy.xsql.block.model.ReferenceBlock;
import com.xy.xsql.block.tsql.core.variable.DeclareVariableConverter;
import org.junit.Assert;
import org.junit.Test;

import java.io.StringWriter;

import static org.junit.Assert.*;

/**
 * Created by xiaoyao9184 on 2017/6/15.
 */
public class AtTimeConverterTest {

    @Test
    public void test() throws Exception {
        ReferenceBlock b = AtTimeConverter.meta();

        StringWriter writer = new ReferenceBlockPrinter()
                .print(b);

        System.out.print(writer);
        Assert.assertEquals(writer.toString(),
                "<AT TIME ZONE> ::=\n" +
                        "inputdate AT TIME ZONE timezone");
    }

}